Setup AK4 Puppi for Nano Run3. Split reco jet configurations into 3 cff files. Move jet tagging switches from nano_cff to AK4 CHS and AK8 cff files. Setup switches to AK4 CHS using Run2 modifiers.
