+++ PhysicsTools/NanoAOD/python/nano_cff.py
@@ -2,7 +2,9 @@ from __future__ import print_function
 import FWCore.ParameterSet.Config as cms
 from PhysicsTools.NanoAOD.common_cff import *
 from PhysicsTools.NanoAOD.nano_eras_cff import *
-from PhysicsTools.NanoAOD.jets_cff import *
+from PhysicsTools.NanoAOD.jetsAK4_CHS_cff import *
+from PhysicsTools.NanoAOD.jetsAK4_Puppi_cff import *
+from PhysicsTools.NanoAOD.jetsAK8_cff import *
 from PhysicsTools.NanoAOD.jetMC_cff import *
 from PhysicsTools.NanoAOD.muons_cff import *
 from PhysicsTools.NanoAOD.taus_cff import *
@@ -34,12 +36,17 @@ nanoMetadata = cms.EDProducer("UniqueStringProducer",
 )
 
 linkedObjects = cms.EDProducer("PATObjectCrossLinker",
-   jets=cms.InputTag("finalJets"),
+   jets=cms.InputTag("finalJetsPuppi"),
    muons=cms.InputTag("finalMuons"),
    electrons=cms.InputTag("finalElectrons"),
    taus=cms.InputTag("finalTaus"),
    photons=cms.InputTag("finalPhotons"),
 )
+run2_nanoAOD_ANY = (
+    run2_miniAOD_80XLegacy | run2_nanoAOD_94X2016 | run2_nanoAOD_94XMiniAODv1 |  run2_nanoAOD_94XMiniAODv2 | run2_nanoAOD_102Xv1 | run2_nanoAOD_106Xv1 | run2_nanoAOD_106Xv2
+)
+# Switch to AK4 CHS jets for Run-2
+run2_nanoAOD_ANY.toModify(linkedObjects, jets="finalJets")
 
 simpleCleanerTable = cms.EDProducer("NanoAODSimpleCrossCleaner",
    name=cms.string("cleanmask"),
@@ -60,20 +67,30 @@ simpleCleanerTable = cms.EDProducer("NanoAODSimpleCrossCleaner",
 
 
 lhcInfoTable = cms.EDProducer("LHCInfoProducer",
-                              precision = cms.int32(10),
+    precision = cms.int32(10),
 )
 
 nanoTableTaskCommon = cms.Task(
-     cms.Task(nanoMetadata), jetTask, jetForMETTask, extraFlagsProducersTask, muonTask, tauTask, boostedTauTask,
-     electronTask , lowPtElectronTask, photonTask,
-     vertexTask, isoTrackTask, jetLepTask,  # must be after all the leptons
-     softActivityTask,
-     cms.Task(linkedObjects),
-     jetTablesTask, muonTablesTask, fsrTablesTask, tauTablesTask, boostedTauTablesTask,
-     electronTablesTask, lowPtElectronTablesTask, photonTablesTask,
-     globalTablesTask, vertexTablesTask, metTablesTask, simpleCleanerTable, extraFlagsTableTask,
-     isoTrackTablesTask,softActivityTablesTask
- )
+    cms.Task(nanoMetadata), 
+    jetPuppiTask, jetPuppiForMETTask, jetAK8Task,
+    extraFlagsProducersTask, muonTask, tauTask, boostedTauTask,
+    electronTask , lowPtElectronTask, photonTask,
+    vertexTask, isoTrackTask, jetAK8LepTask,  # must be after all the leptons
+    softActivityTask,
+    cms.Task(linkedObjects),
+    jetPuppiTablesTask, jetAK8TablesTask,
+    muonTablesTask, fsrTablesTask, tauTablesTask, boostedTauTablesTask,
+    electronTablesTask, lowPtElectronTablesTask, photonTablesTask,
+    globalTablesTask, vertexTablesTask, metTablesTask, simpleCleanerTable, extraFlagsTableTask,
+    isoTrackTablesTask,softActivityTablesTask
+)
+
+# Replace AK4 Puppi with AK4 CHS for Run-2
+_nanoTableTaskCommonRun2 = nanoTableTaskCommon.copy()
+_nanoTableTaskCommonRun2.replace(jetPuppiTask, jetTask)
+_nanoTableTaskCommonRun2.replace(jetPuppiForMETTask, jetForMETTask)
+_nanoTableTaskCommonRun2.replace(jetPuppiTablesTask, jetTablesTask)
+run2_nanoAOD_ANY.toReplaceWith(nanoTableTaskCommon, _nanoTableTaskCommonRun2)
 
 nanoSequenceCommon = cms.Sequence(nanoTableTaskCommon)
 
@@ -82,10 +99,12 @@ nanoSequenceOnlyData = cms.Sequence(cms.Sequence(protonTablesTask) + lhcInfoTabl
 
 nanoSequence = cms.Sequence(nanoSequenceCommon + nanoSequenceOnlyData + nanoSequenceOnlyFullSim)
 
-nanoTableTaskFS = cms.Task(genParticleTask, particleLevelTask, jetMCTask, muonMCTask, electronMCTask, lowPtElectronMCTask, photonMCTask,
-                            tauMCTask, boostedTauMCTask,
-                            metMCTable, ttbarCatMCProducersTask, globalTablesMCTask, cms.Task(btagWeightTable), ttbarCategoryTableTask,
-                            genWeightsTableTask, genVertexTablesTask, genParticleTablesTask, genProtonTablesTask, particleLevelTablesTask)
+nanoTableTaskFS = cms.Task(
+    genParticleTask, particleLevelTask, jetMCTask, muonMCTask, electronMCTask, lowPtElectronMCTask, photonMCTask,
+    tauMCTask, boostedTauMCTask,
+    metMCTable, ttbarCatMCProducersTask, globalTablesMCTask, cms.Task(btagWeightTable), ttbarCategoryTableTask,
+    genWeightsTableTask, genVertexTablesTask, genParticleTablesTask, genProtonTablesTask, particleLevelTablesTask
+)
 
 nanoSequenceFS = cms.Sequence(nanoSequenceCommon + cms.Sequence(nanoTableTaskFS))
 
@@ -131,31 +150,6 @@ def nanoAOD_addBoostedTauIds(process, idsToRun=[]):
 
     return process
 
-
-from PhysicsTools.PatAlgos.tools.jetTools import updateJetCollection
-def nanoAOD_addDeepInfo(process,addDeepBTag,addDeepFlavour):
-    _btagDiscriminators=[]
-    if addDeepBTag:
-        print("Updating process to run DeepCSV btag")
-        _btagDiscriminators += ['pfDeepCSVJetTags:probb','pfDeepCSVJetTags:probbb','pfDeepCSVJetTags:probc']
-    if addDeepFlavour:
-        print("Updating process to run DeepFlavour btag")
-        _btagDiscriminators += ['pfDeepFlavourJetTags:probb','pfDeepFlavourJetTags:probbb','pfDeepFlavourJetTags:problepb','pfDeepFlavourJetTags:probc']
-    if len(_btagDiscriminators)==0: return process
-    print("Will recalculate the following discriminators: "+", ".join(_btagDiscriminators))
-    updateJetCollection(
-               process,
-               jetSource = cms.InputTag('slimmedJets'),
-               jetCorrections = ('AK4PFchs', cms.vstring(['L1FastJet', 'L2Relative', 'L3Absolute','L2L3Residual']), 'None'),
-               btagDiscriminators = _btagDiscriminators,
-               postfix = 'WithDeepInfo',
-           )
-    process.load("Configuration.StandardSequences.MagneticField_cff")
-    process.jetCorrFactorsNano.src="selectedUpdatedPatJetsWithDeepInfo"
-    process.updatedJets.jetSource="selectedUpdatedPatJetsWithDeepInfo"
-    return process
-
-
 from PhysicsTools.PatUtils.tools.runMETCorrectionsAndUncertainties import runMetCorAndUncFromMiniAOD
 def nanoAOD_recalibrateMETs(process,isData):
     # add DeepMETs
@@ -196,18 +190,18 @@ def nanoAOD_recalibrateMETs(process,isData):
         addToProcessAndTask('ak4PuppiJets', ak4PFJets.clone (src = 'puppi', doAreaFastjet = True, jetPtMin = 10.), process, task)
         from PhysicsTools.PatAlgos.tools.jetTools import addJetCollection
         addJetCollection(process,
-                            labelName = 'Puppi',
-                            jetSource = cms.InputTag('ak4PuppiJets'),
-                            algo = 'AK', rParam=0.4,
-                            genJetCollection=cms.InputTag('slimmedGenJets'),
-                            jetCorrections = ('AK4PFPuppi', ['L1FastJet', 'L2Relative', 'L3Absolute','L2L3Residual'], 'None'),
-                            pfCandidates = cms.InputTag('packedPFCandidates'),
-                            pvSource = cms.InputTag('offlineSlimmedPrimaryVertices'),
-                            svSource = cms.InputTag('slimmedSecondaryVertices'),
-                            muSource =cms.InputTag( 'slimmedMuons'),
-                            elSource = cms.InputTag('slimmedElectrons'),
-                            genParticles= cms.InputTag('prunedGenParticles'),
-                            getJetMCFlavour= False
+            labelName = 'Puppi',
+            jetSource = cms.InputTag('ak4PuppiJets'),
+            algo = 'AK', rParam=0.4,
+            genJetCollection=cms.InputTag('slimmedGenJets'),
+            jetCorrections = ('AK4PFPuppi', ['L1FastJet', 'L2Relative', 'L3Absolute','L2L3Residual'], 'None'),
+            pfCandidates = cms.InputTag('packedPFCandidates'),
+            pvSource = cms.InputTag('offlineSlimmedPrimaryVertices'),
+            svSource = cms.InputTag('slimmedSecondaryVertices'),
+            muSource =cms.InputTag( 'slimmedMuons'),
+            elSource = cms.InputTag('slimmedElectrons'),
+            genParticles= cms.InputTag('prunedGenParticles'),
+            getJetMCFlavour= False
         )
 
         process.patJetsPuppi.addGenPartonMatch = cms.bool(False)
@@ -246,52 +240,6 @@ def nanoAOD_activateVID(process):
         modifier.toModify(process.egmPhotonIDs, physicsObjectSrc = "slimmedPhotonsTo106X")
     return process
 
-def nanoAOD_addDeepInfoAK8(process, addDeepBTag, addDeepBoostedJet, addDeepDoubleX, addDeepDoubleXV2, addParticleNet, addParticleNetMass, jecPayload):
-    _btagDiscriminators=[]
-    if addDeepBTag:
-        print("Updating process to run DeepCSV btag to AK8 jets")
-        _btagDiscriminators += ['pfDeepCSVJetTags:probb','pfDeepCSVJetTags:probbb']
-    if addDeepBoostedJet:
-        print("Updating process to run DeepBoostedJet on datasets before 103X")
-        from RecoBTag.ONNXRuntime.pfDeepBoostedJet_cff import _pfDeepBoostedJetTagsAll as pfDeepBoostedJetTagsAll
-        _btagDiscriminators += pfDeepBoostedJetTagsAll
-    if addParticleNet:
-        print("Updating process to run ParticleNet before it's included in MiniAOD")
-        from RecoBTag.ONNXRuntime.pfParticleNet_cff import _pfParticleNetJetTagsAll as pfParticleNetJetTagsAll
-        _btagDiscriminators += pfParticleNetJetTagsAll
-    if addParticleNetMass:
-        from RecoBTag.ONNXRuntime.pfParticleNet_cff import _pfParticleNetMassRegressionOutputs
-        _btagDiscriminators += _pfParticleNetMassRegressionOutputs
-    if addDeepDoubleX:
-        print("Updating process to run DeepDoubleX on datasets before 104X")
-        _btagDiscriminators += ['pfDeepDoubleBvLJetTags:probHbb', \
-            'pfDeepDoubleCvLJetTags:probHcc', \
-            'pfDeepDoubleCvBJetTags:probHcc', \
-            'pfMassIndependentDeepDoubleBvLJetTags:probHbb', 'pfMassIndependentDeepDoubleCvLJetTags:probHcc', 'pfMassIndependentDeepDoubleCvBJetTags:probHcc']
-    if addDeepDoubleXV2:
-        print("Updating process to run DeepDoubleXv2 on datasets before 11X")
-        _btagDiscriminators += [
-            'pfMassIndependentDeepDoubleBvLV2JetTags:probHbb',
-            'pfMassIndependentDeepDoubleCvLV2JetTags:probHcc',
-            'pfMassIndependentDeepDoubleCvBV2JetTags:probHcc'
-            ]
-    if len(_btagDiscriminators)==0: return process
-    print("Will recalculate the following discriminators on AK8 jets: "+", ".join(_btagDiscriminators))
-    updateJetCollection(
-       process,
-       jetSource = cms.InputTag('slimmedJetsAK8'),
-       pvSource = cms.InputTag('offlineSlimmedPrimaryVertices'),
-       svSource = cms.InputTag('slimmedSecondaryVertices'),
-       rParam = 0.8,
-       jetCorrections = (jecPayload.value(), cms.vstring(['L1FastJet', 'L2Relative', 'L3Absolute', 'L2L3Residual']), 'None'),
-       btagDiscriminators = _btagDiscriminators,
-       postfix='AK8WithDeepInfo',
-       printWarning = False
-       )
-    process.jetCorrFactorsAK8.src="selectedUpdatedPatJetsAK8WithDeepInfo"
-    process.updatedJetsAK8.jetSource="selectedUpdatedPatJetsAK8WithDeepInfo"
-    return process
-
 from PhysicsTools.PatUtils.tools.runMETCorrectionsAndUncertainties import runMetCorAndUncFromMiniAOD
 def nanoAOD_runMETfixEE2017(process,isData):
     runMetCorAndUncFromMiniAOD(process,isData=isData,
@@ -304,59 +252,24 @@ def nanoAOD_runMETfixEE2017(process,isData):
 def nanoAOD_customizeCommon(process):
 
     process = nanoAOD_activateVID(process)
-    nanoAOD_addDeepInfo_switch = cms.PSet(
-        nanoAOD_addDeepBTag_switch = cms.untracked.bool(False),
-        nanoAOD_addDeepFlavourTag_switch = cms.untracked.bool(False),
-        )
-    run2_miniAOD_80XLegacy.toModify(nanoAOD_addDeepInfo_switch, nanoAOD_addDeepBTag_switch = cms.untracked.bool(True))
-    for modifier in run2_miniAOD_80XLegacy, run2_nanoAOD_94X2016, run2_nanoAOD_94XMiniAODv1, run2_nanoAOD_94XMiniAODv2:
-        modifier.toModify(nanoAOD_addDeepInfo_switch, nanoAOD_addDeepFlavourTag_switch =  cms.untracked.bool(True))
-    process = nanoAOD_addDeepInfo(process,
-                                  addDeepBTag=nanoAOD_addDeepInfo_switch.nanoAOD_addDeepBTag_switch,
-                                  addDeepFlavour=nanoAOD_addDeepInfo_switch.nanoAOD_addDeepFlavourTag_switch)
-    nanoAOD_addDeepInfoAK8_switch = cms.PSet(
-        nanoAOD_addDeepBTag_switch = cms.untracked.bool(False),
-        nanoAOD_addDeepBoostedJet_switch = cms.untracked.bool(False),
-        nanoAOD_addDeepDoubleX_switch = cms.untracked.bool(False),
-        nanoAOD_addDeepDoubleXV2_switch = cms.untracked.bool(False),
-        nanoAOD_addParticleNet_switch = cms.untracked.bool(False),
-        nanoAOD_addParticleNetMass_switch = cms.untracked.bool(False),
-        jecPayload = cms.untracked.string('AK8PFPuppi')
-        )
-    # deepAK8 should not run on 80X, that contains ak8PFJetsCHS jets
-    run2_miniAOD_80XLegacy.toModify(nanoAOD_addDeepInfoAK8_switch,
-                                    nanoAOD_addDeepBTag_switch = True,
-                                    jecPayload = 'AK8PFchs')
-    # for 94X and 102X samples: needs to run DeepAK8, DeepDoubleX and ParticleNet
-    (run2_nanoAOD_94X2016 | run2_nanoAOD_94XMiniAODv1 | run2_nanoAOD_94XMiniAODv2 | run2_nanoAOD_102Xv1).toModify(
-        nanoAOD_addDeepInfoAK8_switch,
-        nanoAOD_addDeepBoostedJet_switch = True,
-        nanoAOD_addDeepDoubleX_switch = True,
-        nanoAOD_addDeepDoubleXV2_switch = True,
-        nanoAOD_addParticleNet_switch = True,
-        nanoAOD_addParticleNetMass_switch = True,
-    )
-    # for 106Xv1: only needs to run ParticleNet and DDXV2; DeepAK8, DeepDoubleX are already in MiniAOD
-    run2_nanoAOD_106Xv1.toModify(
-        nanoAOD_addDeepInfoAK8_switch,
-        nanoAOD_addDeepDoubleXV2_switch = True,
-        nanoAOD_addParticleNet_switch = True,
-        nanoAOD_addParticleNetMass_switch = True,
+
+    # This function is defined in jetsAK4_CHS_cff.py
+    process = nanoAOD_addDeepInfoAK4CHS(process,
+        addDeepBTag=nanoAOD_addDeepInfoAK4CHS_switch.nanoAOD_addDeepBTag_switch,
+        addDeepFlavour=nanoAOD_addDeepInfoAK4CHS_switch.nanoAOD_addDeepFlavourTag_switch
     )
 
-    run2_nanoAOD_106Xv2.toModify(
-         nanoAOD_addDeepInfoAK8_switch,
-         nanoAOD_addParticleNetMass_switch = True,
-    )
-
+    # This function is defined in jetsAK8_cff.py
     process = nanoAOD_addDeepInfoAK8(process,
-                                     addDeepBTag=nanoAOD_addDeepInfoAK8_switch.nanoAOD_addDeepBTag_switch,
-                                     addDeepBoostedJet=nanoAOD_addDeepInfoAK8_switch.nanoAOD_addDeepBoostedJet_switch,
-                                     addDeepDoubleX=nanoAOD_addDeepInfoAK8_switch.nanoAOD_addDeepDoubleX_switch,
-                                     addDeepDoubleXV2=nanoAOD_addDeepInfoAK8_switch.nanoAOD_addDeepDoubleXV2_switch,
-                                     addParticleNet=nanoAOD_addDeepInfoAK8_switch.nanoAOD_addParticleNet_switch,
-                                     addParticleNetMass=nanoAOD_addDeepInfoAK8_switch.nanoAOD_addParticleNetMass_switch,
-                                     jecPayload=nanoAOD_addDeepInfoAK8_switch.jecPayload)
+        addDeepBTag=nanoAOD_addDeepInfoAK8_switch.nanoAOD_addDeepBTag_switch,
+        addDeepBoostedJet=nanoAOD_addDeepInfoAK8_switch.nanoAOD_addDeepBoostedJet_switch,
+        addDeepDoubleX=nanoAOD_addDeepInfoAK8_switch.nanoAOD_addDeepDoubleX_switch,
+        addDeepDoubleXV2=nanoAOD_addDeepInfoAK8_switch.nanoAOD_addDeepDoubleXV2_switch,
+        addParticleNet=nanoAOD_addDeepInfoAK8_switch.nanoAOD_addParticleNet_switch,
+        addParticleNetMass=nanoAOD_addDeepInfoAK8_switch.nanoAOD_addParticleNetMass_switch,
+        jecPayload=nanoAOD_addDeepInfoAK8_switch.jecPayload
+    )
+
     nanoAOD_tau_switch = cms.PSet(
         idsToAdd = cms.vstring()
     )
@@ -372,7 +285,6 @@ def nanoAOD_customizeCommon(process):
 
 def nanoAOD_customizeData(process):
     process = nanoAOD_customizeCommon(process)
-
     for modifier in run2_miniAOD_80XLegacy,run2_nanoAOD_94X2016,run2_nanoAOD_94XMiniAODv1,run2_nanoAOD_94XMiniAODv2,run2_nanoAOD_102Xv1,run2_nanoAOD_106Xv1:
         modifier.toModify(process, lambda p: nanoAOD_recalibrateMETs(p,isData=True))
     for modifier in run2_nanoAOD_94XMiniAODv1, run2_nanoAOD_94XMiniAODv2:
